+++ tests/testthat/test-make_frames.r
@@ -18,16 +18,17 @@ pos1 <- structure(list(
   )
 ), row.names = 4:7, class = "data.frame")
 
-# make preview image
+# Check output with default background_ylim, background_xlim, and bg_map
+# make a preview image
 temp_dir <- tempdir()
 make_frames(pos1, out_dir = temp_dir, preview = TRUE)
 
-# Actual file sizes
-img_file <- file.path(temp_dir, "1.png")
-img_size <- file.info(img_file)$size
+# Expected file size
+size_should_be1 <- 30919
 
-# Expected file sizes
-size_should_be <- 30919
+# Actual file sizes
+img_file1 <- file.path(temp_dir, "1.png")
+img_size1 <- file.info(img_file1)$size
 
 # Clean up
 unlink(list.files(temp_dir,
@@ -35,8 +36,99 @@ unlink(list.files(temp_dir,
   include.dirs = TRUE
 ), recursive = TRUE)
 
+# Check output with default background_ylim, background_xlim, but not bg_map
+# make a preview image
+data(great_lakes_polygon)
+
+temp_dir <- tempdir()
+make_frames(pos1,
+  out_dir = temp_dir, preview = TRUE,
+  bg_map = great_lakes_polygon
+)
+
+# Expected file size
+size_should_be2 <- 23832
+
+# Actual file sizes
+img_file2 <- file.path(temp_dir, "1.png")
+img_size2 <- file.info(img_file2)$size
+
+# Clean up
+unlink(list.files(temp_dir,
+  full.names = TRUE, recursive = TRUE,
+  include.dirs = TRUE
+), recursive = TRUE)
+
+# Check output with specified background_ylim, background_xlim, and bg_map
+# make a preview image
+data(great_lakes_polygon)
+
+temp_dir <- tempdir()
+make_frames(pos1,
+  out_dir = temp_dir,
+  background_ylim = c(42, 47),
+  background_xlim = c(-90, -78),
+  preview = TRUE,
+  bg_map = great_lakes_polygon
+)
+
+# Expected file size
+size_should_be3 <- 32220
+
+# Actual file sizes
+img_file3 <- file.path(temp_dir, "1.png")
+img_size3 <- file.info(img_file3)$size
+
+# Clean up
+unlink(list.files(temp_dir,
+  full.names = TRUE, recursive = TRUE,
+  include.dirs = TRUE
+), recursive = TRUE)
+
+# Check output with default background_ylim, background_xlim, but not bg_map
+# bg_map input class = terra::spatVector
+# make a preview image
+
+sv_poly <- terra::vect(great_lakes_polygon)
+
+temp_dir <- tempdir()
+make_frames(pos1,
+  out_dir = temp_dir,
+  preview = TRUE,
+  bg_map = sv_poly
+)
+
+# Expected file size
+size_should_be4 <- 23832
+
+# Actual file sizes
+img_file4 <- file.path(temp_dir, "1.png")
+img_size4 <- file.info(img_file4)$size
+
+# Clean up
+unlink(list.files(temp_dir,
+  full.names = TRUE, recursive = TRUE,
+  include.dirs = TRUE
+), recursive = TRUE)
+
+
 # Testing file size results
-test_that("making preview image expected result", {
+test_that("Expected result when background lims and map not supplied", {
   # Check if expected and actual file sizes
-  expect_equal(img_size, size_should_be)
+  expect_equal(img_size1, size_should_be1)
+})
+
+test_that("Expected result when map but not background lims supplied", {
+  # Check if expected and actual file sizes
+  expect_equal(img_size2, size_should_be2)
+})
+
+test_that("Expected result when map and background lims supplied", {
+  # Check if expected and actual file sizes
+  expect_equal(img_size3, size_should_be3)
+})
+
+test_that("Expected result when map is spatVector, bg lims not supplied", {
+  # Check if expected and actual file sizes
+  expect_equal(img_size4, size_should_be4)
 })