+++ NEWS.md
@@ -1,3 +1,4 @@
+----
 
 # glatos 0.8.0 (dev)
 
@@ -13,6 +14,13 @@
 
 
 ### New features
+
+#### 2024-01-19
+
+- In `make_frames()`, allow `terra::SpatVector` input for background map 
+  (`bg_map` arg).
+    - add test for `terra::SpatVector` input
+
 
 #### 2023-08-30
 
@@ -65,6 +73,20 @@
 - Add new function `read_vdat_csv()` to read data exported from VRL or VDAT 
   using VDAT.exe (e.g., using `vdat_convert()`).
 
+
+----
+
+# glatos 0.7.1 (2024-01-19)
+
+
+### Bug fixes and minor changes
+
+
+- Fix bug introduced in glatos 0.7.0 where x and y limits were switched 
+  when `bg_map` was supplied.
+    - Add test for various inputs of `background_xlim`, `background_ylim`, and 
+      `bg_map` to function `make_frames`.
+  
 
 ----